NetworkScanner/requirements.txt
PyQt6>=6.4.0
scapy>=2.4.5
python-nmap>=0.7.1
networkx>=2.8.0
matplotlib>=3.5.0
netifaces>=0.11.0
psutil>=5.9.0
reportlab>=3.6.12
pillow>=9.3.0
sqlalchemy>=1.4.41
